Renamed old comps
