+++ packages/commonwealth/client/scripts/views/components/component_kit/cw_toggle_old.scss
