+++ packages/commonwealth/client/scripts/views/pages/CommunityManagement/Integrations/Directory/Directory.tsx
@@ -11,7 +11,7 @@ import {
 import { useFetchNodesQuery } from 'state/api/nodes';
 import { CWLabel } from 'views/components/component_kit/cw_label';
 import { CWText } from 'views/components/component_kit/cw_text';
-import { CWToggle } from 'views/components/component_kit/cw_toggle';
+import { CWToggle } from 'views/components/component_kit/cw_toggle_old';
 import { CWButton } from 'views/components/component_kit/new_designs/CWButton';
 import { CWTooltip } from 'views/components/component_kit/new_designs/CWTooltip';
 import { CWTypeaheadSelectList } from 'views/components/component_kit/new_designs/CWTypeaheadSelectList';
