+++ packages/commonwealth/client/scripts/views/pages/NotificationSettings/PushNotificationsToggle.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback } from 'react';
 // eslint-disable-next-line max-len
-import { CWToggle } from 'views/components/component_kit/cw_toggle';
+import { CWToggle } from 'views/components/component_kit/cw_toggle_old';
 import {
   SubscriptionPrefType,
   useSubscriptionPreferenceSetting,
